+++ apps/extension/cypress/stubs/flipScore.tsx
@@ -1,0 +1,2 @@
+const flipScore = ({ value }: { value: number }) => <span>{value}</span>;
+export default flipScore;