+++ apps/extension/cypress/stubs/tabAssignSelect.tsx
@@ -1,0 +1,2 @@
+const tabAssignSelect = () => null;
+export default tabAssignSelect;