+++ apps/extension/cypress/support/commands.ts
@@ -1,0 +1,2 @@
+// Custom Cypress commands go here.
+// See https://on.cypress.io/custom-commands